Made the score make sense #20
This addresses issue #20, making the score more intuitive.

Score now increases with each kill. Each enemy has a score granted on death. And the score value on the HUD increases linearly within a set timeframe!

This also helps with issue #21, giving some feedback when killing, but not completely resolving that issue.